Refactoring Shape
* Shoes3/shoes3#333
