+++ shoes/types/color.h
@@ -16,10 +16,6 @@ extern shoes_app _shoes_app;
 extern void shoes_pattern_gradient(shoes_pattern *pattern, VALUE r1, VALUE r2, VALUE attr);
 
 // native forward declarations
-/*extern SHOES_CONTROL_REF shoes_native_spinner(VALUE self, shoes_canvas *canvas, shoes_place *place, VALUE attr, char *msg);
-extern void shoes_native_spinner_start(SHOES_CONTROL_REF ref);
-extern void shoes_native_spinner_stop(SHOES_CONTROL_REF ref);
-extern gboolean shoes_native_spinner_started(SHOES_CONTROL_REF ref);*/
 
 VALUE cColor, cColors;
